requirements.txt
pandas
dotenv
websocket-client
threading
